Initialize repository for 2022 season
Erased all files and generated a new 2022 FRC Command-Based project from the WPILib project generator.
