+++ .gitignore
@@ -1,4 +1,5 @@
-# Created by https://www.gitignore.io/api/c++,java,linux,macos,gradle,windows,visualstudiocode
+# This gitignore has been specially created by the WPILib team.
+# If you remove items from this file, intellisense might break.
 
 ### C++ ###
 # Prerequisites
@@ -151,11 +152,11 @@ gradle-app.setting
 # gradle/wrapper/gradle-wrapper.properties
 
 # # VS Code Specific Java Settings
+# DO NOT REMOVE .classpath and .project
 .classpath
 .project
 .settings/
 bin/
-imgui.ini
 
-
-# End of https://www.gitignore.io/api/c++,java,linux,macos,gradle,windows,visualstudiocode
+# Simulation GUI and other tools window save file
+*-window.json
