+++ .wpilib/wpilib_preferences.json
@@ -1,6 +1,6 @@
 {
     "enableCppIntellisense": false,
     "currentLanguage": "java",
-    "projectYear": "2021",
+    "projectYear": "2022",
     "teamNumber": 7221
 }
