+++ src/main/java/frc/robot/Constants.java
@@ -12,35 +12,4 @@ package frc.robot;
  * <p>It is advised to statically import this class (or one of its inner classes) wherever the
  * constants are needed, to reduce verbosity.
  */
-public final class Constants {
-
-  /** Talon CAN Bus ID's for various subsystems */
-  public final class Talon {
-
-    /** Talon CAN Bus ID for {@link Drivetrain} subsystem */
-    public final class Drivetrain {
-      /** CAN Bus ID of primary left drive Talon */
-      public final static int LEFT_MAIN = 3;
-      /** CAN Bus ID of follower left drive Talon or Victor */
-      public final static int LEFT_FOLLOW = 53;
-      /** CAN Bus ID of primary right drive Talon */
-      public final static int RIGHT_MAIN = 1;
-      /** CAN Bus ID of follower right drive Talon or Victor */
-      public final static int RIGHT_FOLLOW = 51;
-      /** Left side Talons are inverted */
-      public final static boolean LEFT_INVERTED = false;
-      /** Right side Talons are inverted */
-      public final static boolean RIGHT_INVERTED = false;
-    }
-
-  }
-
-  /** Joystick USB ports */
-  public final class Joystick {
-    /** USB port for driver controller */
-    public final static int DRIVER = 0;
-    /** USB port for support controller */
-    public final static int SUPPORT = 1;
-  }
-
-}
+public final class Constants {}
